Moving DTOs folder from WebApi to Infrastructure
