--- HospitalManagementAndAppointmentSystem/DTOs/PatientLoginDto.cs
@@ -1,13 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace HospitalManagementAndAppointmentSystem.DTOs
-{
-    public class PatientLoginDto
-    {
-        [Required]
-        [EmailAddress]
-        public string Email {  get; set; }
-        [Required]
-        public string Password { get; set; }
-    }
-}
