+++ HospitalManagementAndAppointmentSystem/HospitalManagementAndAppointmentSystem.csproj
@@ -23,9 +23,4 @@
 		<ProjectReference Include="..\Infrastructure\Infrastructure.csproj" />
 	</ItemGroup>
 
-	<ItemGroup>
-		<Folder Include="Service\" />
-		<Folder Include="DTOs\" />
-	</ItemGroup>
-
 </Project>
