+++ Infrastructure/DTOs/HelpDeskLoginDto.cs
@@ -1,13 +1,13 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace HospitalManagementAndAppointmentSystem.DTOs
+namespace Infrastructure.DTOs
 {
-    public class DoctorLoginDto
+    public class HelpDeskLoginDto
     {
         [Required]
         [EmailAddress]
         public string Email { get; set; }
         [Required]
-        public string Password {  get; set; }
+        public string Password { get; set; }
     }
 }
